Initial implementation of transposition table
There are some parts of the code that are commented out. These parts are left
commented so that the decision can be made later whether or not to store the
transpositions (with the trade-off of using 8 times more memory) or check each
time (which obviously hurts the runtime).

Benchmarking needs to be done whether it's worth not passing the
DomineeringState as const (and thus obfuscating the code) and modifying it in
such a way that it goes back to the original configuration at the end of
TranspositionTable::insert.

Here are some TODOs for the transposition table:
* reading in from file for the initial set of data to use
* remove code commented by making the decision described above
* decide whether to use const or not for TranspositionTable::insert
* implement removing of elements so that table doesn't exceed memory limit
  - this could result in change of data structure (since additional
    information is needed to determine which element to remove)
  - the maximum size needs to be defined in this regard
* implement feature to export current table to file