[BUGFIX] incorrect hashing logic
This change fixes the bug where incorrect values are hashed by not considering
the location of the empty grid, resulting in more collisions, thus, very slow
operations on unordered_map. Moreover, the previous implementation does not
reflect the current state of the board accurately, which results in incorrect
values being stored in the transposition table.