Use difference in reserved spot for the eval function
This commit makes the change of looking at only the difference in the number
of "reserved" spots, where the opponent's domino cannot be placed.

This deprecates the EvalTakeAway class, which was created to see how much
improvement the team can get by placing a domino at a certain location by
looking at the number of opponent's potential placement the placement takes
away.